Refactoring directory structure
