+++ test/core/sup_tree_core/periodic_log_writer_test.exs
@@ -4,8 +4,8 @@ defmodule AntikytheraCore.PeriodicLogWriterTest do
   @dir Path.join([__DIR__, "..", "..", "..", "_build", "test", "log", "antikythera"]) |> Path.expand()
 
   test "should exists" do
-    assert Process.whereis(PeriodicLogWriter.Reduction) != nil
-    send(PeriodicLogWriter.Reduction, :timeout)
+    assert Process.whereis(AntikytheraCore.PeriodicLog.ReductionBuilder) != nil
+    send(AntikytheraCore.PeriodicLog.ReductionBuilder, :timeout)
     assert File.exists?(Path.join(@dir, "reduction.log.gz"))
   end
 end
